video-studio/requirements.txt
fastapi
uvicorn
openai
moviepy
pillow
matplotlib
gtts
python-multipart
python-dotenv
